// client/src/models/IPersonalInformation.ts
export interface IPersonalInformation {
    id: number;
    gender:  string;
    lastName: string;
    name: string;
    patronymic: string;
    dateOfBirth: string;
    placeOfBirth: string;
    placeOfRegistration: string;
    placeOfResidence: string;
    phone: string;
    citizenship: string;
    education: string;
    maritalStatus: string;
    placeOfWork: string;
    militaryDuty: string;
    criminalRecord: string;
    passport: string;
    otherData: string;
    
}